Переносит использование access_token лесты из localStorage в redux store
