+++ src/entities/User/model/selectors/index.ts
@@ -6,3 +6,4 @@ export const getLoggedInStatus = (state: StateSchema) => state?.user?.isLoggedIn
 export const getCurrentUserError = (state: StateSchema) => state?.user?.error || '';
 export const getFullUserState = (state: StateSchema) => state?.user || null;
 export const getTokenUpdateStatus = (state: StateSchema) => state?.user?.isTokenRefreshing || false;
+export const getLestaAccessToken = (state: StateSchema) => state?.user?.authData?.lestaData?.access_token || null;
