+++ src/entities/User/model/slice/userSlice.ts
@@ -1,9 +1,9 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { LOCAL_STORAGE_LESTA, LOCAL_STORAGE_USER_KEY } from 'shared/consts/localstorage';
+import { LOCAL_STORAGE_USER_KEY } from 'shared/consts/localstorage';
 import { authByLestaOpenID } from 'features/AuthUser/index';
-import { logoutUser, refreshLestaToken } from 'entities/User/index';
-import { User, UserSchema } from '../types/user';
+import { logoutUser } from '../services/logoutUser/logoutUser';
 import { checkUserAuth } from '../services/checkUserAuth/checkUserAuth';
+import { User, UserSchema } from '../types/user';
 
 const initialState: UserSchema = {
   isLoggedIn: !!localStorage.getItem(LOCAL_STORAGE_USER_KEY),
@@ -35,38 +35,24 @@ export const userSlice = createSlice({
         state.error = '';
         state.isLoading = true;
         state.isInitiated = true;
+        state.isTokenRefreshing = true;
       })
       .addCase(checkUserAuth.rejected, (state, { payload }) => {
         state.error = payload;
         state.isLoading = false;
         state.isLoggedIn = false;
+        state.isTokenRefreshing = false;
         localStorage.clear();
       })
       .addCase(checkUserAuth.fulfilled, (state, { payload }) => {
         state.isLoading = false;
         state.isLoggedIn = true;
+        state.isTokenRefreshing = false;
         state.authData = payload;
       })
       .addCase(authByLestaOpenID.fulfilled, (state, { payload }) => {
         state.isLoggedIn = true;
         state.authData = payload;
-      })
-      .addCase(refreshLestaToken.pending, (state) => {
-        state.isTokenRefreshing = true;
-      })
-      .addCase(refreshLestaToken.fulfilled, (state, { payload }) => {
-        state.isTokenRefreshing = false;
-
-        localStorage.setItem(LOCAL_STORAGE_LESTA.TOKEN, JSON.stringify(payload.access_token));
-        localStorage.setItem(LOCAL_STORAGE_LESTA.EXPIRES_AT, JSON.stringify(payload.expires_at));
-      })
-      .addCase(refreshLestaToken.rejected, (state, { payload }) => {
-        state.isTokenRefreshing = false;
-
-        if (payload === 407) {
-          state.isLoggedIn = false;
-          localStorage.clear();
-        }
       })
       .addCase(logoutUser.fulfilled, (state) => {
         state.isLoggedIn = false;
