+++ src/features/AuthUser/model/services/authByLestaOpenID/authByLestaOpenID.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { ThunkConfig } from 'app/providers/StoreProvider/index';
 import { SERVER_ERROR_MESSAGE } from 'shared/consts/global';
-import { LOCAL_STORAGE_LESTA, LOCAL_STORAGE_USER_KEY } from 'shared/consts/localstorage';
+import { LOCAL_STORAGE_USER_KEY } from 'shared/consts/localstorage';
 import { User, UserOpenID } from 'entities/User';
 import { openIdDTOConverter } from '../../dto/openIdDTOConverter';
 
@@ -33,9 +33,6 @@ export const authByLestaOpenID = createAsyncThunk<User, ThunkProps, ThunkConfig<
       // прогоняем ответ через DTO converter
       const convertedUserData = openIdDTOConverter(currentUserData.data.userData);
       localStorage.setItem(LOCAL_STORAGE_USER_KEY, JSON.stringify(convertedUserData));
-      localStorage.setItem(LOCAL_STORAGE_LESTA.TOKEN, JSON.stringify(ThunkProps.access_token));
-      localStorage
-        .setItem(LOCAL_STORAGE_LESTA.EXPIRES_AT, JSON.stringify(ThunkProps.expires_at));
 
       extra.navigate(`/user/${ThunkProps.account_id}`);
